requirements.txt
python-dotenv
google-cloud-firestore
flask
requests
google-cloud-aiplatform
PyYAML
